Connected python crawler to elasticsearch
