+++ python_crawler/requirements.txt
@@ -2,3 +2,4 @@ beautifulsoup4
 requests
 sqlalchemy
 psycopg2
+elasticsearch
