--- python_crawler/sample_output.txt
@@ -1,8 +1,0 @@
-{
-    "https://en.wikipedia.org/wiki/Crawler": {
-        "text": "\n\n\n\nCrawler - Wikipedia\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nCrawler\n\nFrom Wikipedia, the free encyclopedia\n\n\n\nJump to navigation\nJump to search\n\n\n\nLook up crawler in Wiktionary, the free dictionary.\n\nCrawler may refer to:\n\nWeb crawler, a computer program that gathers and categorizes information on the World Wide Web\nA first-instar nymph of a scale insect that has legs and walks around before it attaches itself and becomes stationary\nCrawler (BEAM) in robotics\nA type of crane on tracks\n\"Crawlers\" (Into the Dark), an episode of the second season of Into the Dark\nAs a proper name:\nCrawler (band), a British rock band\nA fictional creature in the video game Fable III\nA fictional creature in the movie The Descent\nSee also[edit]\nCrawl (disambiguation)\nBottom crawler, an underwater exploration and recovery vehicle\nCrawler-transporter, a large tracked vehicle used by NASA to transport spacecraft\nNightcrawler (Lumbricus terrestris), an annelid worm also called \"common earthworm\" and \"dew worm\"\n\u2039\u00a0The template below (Disambiguation) is being considered for merging. See templates for discussion to help reach a consensus.\u00a0\u203a\nDisambiguation page providing links to topics that could be referred to by the same search term\n\n This disambiguation page lists  articles associated with the title Crawler.  If an internal link led you here, you may wish to change the link to point directly to the intended article. \n\n\n\n\n\nRetrieved from \"https://en.wikipedia.org/w/index.php?title=Crawler&oldid=941555361\"\nCategories: Disambiguation pagesHidden categories: Disambiguation pages with short descriptionsShort description is different from WikidataAll article disambiguation pagesAll disambiguation pages\n\n\n\n\n\n\nNavigation menu\n\n\n\n\nPersonal tools\n\n\nNot logged inTalkContributionsCreate accountLog in\n\n\n\n\n\n\nNamespaces\n\n\nArticleTalk\n\n\n\n\n\n\nVariants\n\n\n\n\n\n\n\n\n\n\nViews\n\n\nReadEditView history\n\n\n\n\n\n\nMore\n\n\n\n\n\n\n\nSearch\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nNavigation\n\n\nMain pageContentsCurrent eventsRandom articleAbout WikipediaContact usDonate\n\n\n\n\n\nContribute\n\n\nHelpLearn to editCommunity portalRecent changesUpload file\n\n\n\n\n\nTools\n\n\nWhat links hereRelated changesUpload fileSpecial pagesPermanent linkPage informationCite this pageWikidata item\n\n\n\n\n\nPrint/export\n\n\nDownload as PDFPrintable version\n\n\n\n\n\nLanguages\n\n\n\ud55c\uad6d\uc5b4Nederlands\nEdit links\n\n\n\n\n\n\n This page was last edited on 19 February 2020, at 07:57\u00a0(UTC).\nText is available under the Creative Commons Attribution-ShareAlike License;\nadditional terms may apply.  By using this site, you agree to the Terms of Use and Privacy Policy. Wikipedia\u00ae is a registered trademark of the Wikimedia Foundation, Inc., a non-profit organization.\n\n\nPrivacy policy\nAbout Wikipedia\nDisclaimers\nContact Wikipedia\nMobile view\nDevelopers\nStatistics\nCookie statement\n\n\n\n\n\n\n\n\n\n\n",
-        "structured_data": [
-            "{\"@context\":\"https:\\/\\/schema.org\",\"@type\":\"Article\",\"name\":\"Crawler\",\"url\":\"https:\\/\\/en.wikipedia.org\\/wiki\\/Crawler\",\"sameAs\":\"http:\\/\\/www.wikidata.org\\/entity\\/Q5182958\",\"mainEntity\":\"http:\\/\\/www.wikidata.org\\/entity\\/Q5182958\",\"author\":{\"@type\":\"Organization\",\"name\":\"Contributors to Wikimedia projects\"},\"publisher\":{\"@type\":\"Organization\",\"name\":\"Wikimedia Foundation, Inc.\",\"logo\":{\"@type\":\"ImageObject\",\"url\":\"https:\\/\\/www.wikimedia.org\\/static\\/images\\/wmf-hor-googpub.png\"}},\"datePublished\":\"2003-10-24T19:29:56Z\",\"dateModified\":\"2020-02-19T07:57:32Z\",\"headline\":\"Wikimedia disambiguation page\"}"
-        ]
-    }
-}
